+++ app/models/adjustable_schema/relationship.rb
@@ -131,20 +131,5 @@ module AdjustableSchema
 		end
 
 		def abstract? = not (source or target)
-
-		# HACK
-		# Using polymorphic associations in combination with single table inheritance (STI) is
-		# a little tricky. In order for the associations to work as expected, ensure that you
-		# store the base model for the STI models in the type column of the polymorphic
-		# association.
-		# https://api.rubyonrails.org/classes/ActiveRecord/Associations/ClassMethods.html#module-ActiveRecord::Associations::ClassMethods-label-Polymorphic+Associations
-		reflections
-				.values
-				.select { it.options[:polymorphic] }
-				.each do |reflection|
-					define_method "#{reflection.name}_type=" do |type|
-						super type && type.to_s.classify.constantize.base_class.to_s
-					end
-				end
 	end
 end